requirements.txt
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.4
daphne==4.2.0
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
oauthlib==3.3.0
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-decouple==3.8
requests==2.32.4
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
urllib3==2.5.0
zope.interface==7.2
